+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/club/core/dto/request/ClubRequestDto.java
@@ -1,8 +1,8 @@
 package com.mobble.mobbleserver.infrastructure.web.club.core.dto.request;
 
+import com.mobble.mobbleserver.domain.club.ageGroup.AgeGroupType;
 import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.infrastructure.web.address.dto.request.AddressRequestDto;
-import com.mobble.mobbleserver.refactor.club.ageGroup.entity.AgeGroupType;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import jakarta.validation.constraints.*;
