+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/club/core/dto/response/ClubResponseDto.java
@@ -1,10 +1,10 @@
 package com.mobble.mobbleserver.infrastructure.web.club.core.dto.response;
 
+import com.mobble.mobbleserver.domain.address.Address;
+import com.mobble.mobbleserver.domain.club.ageGroup.AgeGroupType;
 import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.infrastructure.persistence.club.core.projection.ClubLikeInfoDto;
 import com.mobble.mobbleserver.infrastructure.web.address.dto.response.AddressResponseDto;
-import com.mobble.mobbleserver.domain.address.Address;
-import com.mobble.mobbleserver.refactor.club.ageGroup.entity.AgeGroupType;
 import com.mobble.mobbleserver.infrastructure.web.ground.dto.response.GroundResponseDto;
 
 import java.util.List;
